requirements.txt
fastapi==0.110.3
uvicorn==0.29.0
crewai==0.130.0
crewai-tools==0.47.1
openai==1.30.5
python-dotenv==1.0.1
pypdf==4.2.0
SQLAlchemy==2.0.30